Finished a LOT of the encode/decode numbers
Hey guys.. I finished all the numbers. So it works AMAZINGLY well on my computer, but thats only because my computer can cout any ASCII char so its actually really cool. But unfortunately, that is not the case on xCode. For this program to work successfully on xCode it requires more updates which we need to talk about. Right now that is the only issue; otherwise we are pretty much done unless you guys want to make any major changes. 
Happy Thanksgiving!