Fixed the issue. Finished commenting completely
I fixed the extended ASCII characters problem that we had in XCode. I made the changes to the inputted word relatively less in the ASCII table. This seems to have fixed that issue. The only remaining issue (not really an issue, but an annoyance) is an open issue in github.